Fixing lint errors

+++ internal/vm/hcs/processor.go
@@ -3,6 +3,6 @@
 package hcs
 
 func (uvmb *utilityVMBuilder) SetProcessorCount(count uint32) error {
-	uvmb.doc.VirtualMachine.ComputeTopology.Processor.Count = int32(count)
+	uvmb.doc.VirtualMachine.ComputeTopology.Processor.Count = count
 	return nil
 }
